requirements.txt
cycler==0.10.0
future==0.18.2
imageio==2.13.5
imageio-ffmpeg==0.4.5
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
pygame==2.0.2
pyparsing==3.0.1
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
tk==0.1.0
